scripts/train.py
#!/usr/bin/env python
"""
Training script for the CSIRO Pasture Biomass Estimation competition.
"""

import argparse
import yaml
import torch
from pathlib import Path

from src.utils.seed import set_seed
from src.utils.logger import setup_logger


def parse_args():
    """Parse command line arguments."""
    parser = argparse.ArgumentParser(description='Train model for pasture biomass estimation')
    parser.add_argument('--config', type=str, default='config/config.yaml',
                        help='Path to configuration file')
    parser.add_argument('--gpu', type=int, default=0,
                        help='GPU device ID')
    return parser.parse_args()


def load_config(config_path: str) -> dict:
    """Load configuration from YAML file."""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config


def main():
    """Main training function."""
    args = parse_args()
    
    # Load configuration
    config = load_config(args.config)
    
    # Set random seed
    set_seed(config.get('seed', 42))
    
    # Setup logger
    logger = setup_logger(log_dir=config['paths']['log_dir'])
    logger.info("Starting training process...")
    logger.info(f"Configuration: {config}")
    
    # Set device
    device = torch.device(f"cuda:{args.gpu}" if torch.cuda.is_available() and config['device'] == 'cuda' else 'cpu')
    logger.info(f"Using device: {device}")
    
    # TODO: Implement data loading
    # TODO: Implement model initialization
    # TODO: Implement training loop
    # TODO: Implement validation
    # TODO: Implement model saving
    
    logger.info("Training completed!")


if __name__ == '__main__':
    main()

