requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning & Deep Learning
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.8.0
scikit-learn>=1.0.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.1.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Kaggle API
kaggle>=1.5.12

# Image Processing
scikit-image>=0.19.0

# Model Training Utilities
wandb>=0.12.0  # Weights & Biases for experiment tracking
tensorboard>=2.8.0

# Development Tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: For advanced models
timm>=0.6.0  # PyTorch Image Models
transformers>=4.20.0  # Hugging Face Transformers

